- fixed player stats (str, int, stam, etc)
- fixed  target trigger and proc spells spells applying to wrong targets
- fixed stacking auras subtracting more than they added
-> fixed Spirit Tap, Shadow Weaving and probably others

+++ Services/WCell.RealmServer/Entities/Character.Fields.cs
@@ -1049,20 +1049,20 @@ namespace WCell.RealmServer.Entities
 			this.UpdateAllDamages();
 		}
 
-		private void ModDamageBonusPct(DamageSchool school, float delta)
+		private void ModDamageBonusPct(DamageSchool school, int delta)
 		{
 			if (delta == 0)
 			{
 				return;
 			}
 			var field = PlayerFields.MOD_DAMAGE_DONE_PCT + (int)school;
-			SetFloat(field, GetFloat(field) + delta);
+			SetInt32(field, GetInt32(field) + delta);
 		}
 
 		/// <summary>
 		/// Adds/Removes a percent modifier to all of the given damage schools
 		/// </summary>
-		public void ModDamageBonusPct(uint[] schools, float delta)
+		public void ModDamageBonusPct(uint[] schools, int delta)
 		{
 			foreach (var school in schools)
 			{
